requirements.txt
fastapi==0.61.0
python-dotenv
pydantic
bs4
ojson
requests
uvicorn